refactor: move search to components
