+++ site/src/components/search_bar/helpers.ts
