+++ site/src/components/search_bar/index.ts
@@ -1,0 +1,3 @@
+export { SearchBar } from './search_bar'
+
+export type { SearchBarProps } from './search_bar'
