+++ site/src/components/search_bar/search_bar.stories.tsx
