+++ site/src/features/pages/home/components/page.tsx
@@ -18,7 +18,7 @@ import { Notification } from '~/components/notification'
 import { StationList } from '~/components/station_list'
 import { useToast } from '~/components/toast'
 import { ToggleButton } from '~/components/toggle_button'
-import { SearchBar } from '~/features/search'
+import { SearchBar } from '~/components/search_bar'
 import { useClientStore } from '~/hooks/use_client_store'
 import { useFavorites } from '~/hooks/use_favorites'
 import { useLocale, useTranslations } from '~/i18n'
