--- site/src/features/search/index.ts
@@ -1,3 +1,0 @@
-export { SearchBar } from './components/search_bar'
-
-export type { SearchBarProps } from './components/search_bar'
